bot/requirements.txt
discord.py==2.2.3
httpx==0.24.1
python-dotenv==1.0.0
